+++ core/components/romanesco/elements/snippets/06_formulas/f_formblocks/fbloadassets.snippet.php
@@ -7,9 +7,11 @@ $uploadFile = $modx->getOption('uploadFile', $scriptProperties, 0);
 
 // Load Semantic UI form component separately
 $modx->regClientCSS($assetsPathDist . '/components/form.min.css');
+$modx->regClientCSS($assetsPathDist . '/components/calendar.min.css');
 
 // Load FormBlocks JS in footer
 $modx->regClientScript($assetsPathDist . '/components/form.min.js');
+$modx->regClientScript($assetsPathDist . '/components/calendar.min.js');
 $modx->regClientScript($assetsPathJS . '/formblocks.js');
 
 // Load additional assets for file upload field, if present